Copy/paste/delete of lists of folders and objects now updates the LuceneIndex properly.
Also extended the IndexCommand to allow signalling when the LuceneActor should reload the indexable from the database - and when not (in case of already running transactions with unsaved instances, it is not useful trying to load an unsaved instance from the database. But when re-indexing at startup, this has to be done).