Refactored LuceneService - and fixed it.
Background: LuceneService was traditionally called to update the index when indexable objects changed. This becomes problematic when an Exception causes an automatic rollback of the database - and not of the Lucene index. Such cases are rare, as the indexing almost always happens at after the main action (updating, cloning, removing an object), but it's theoretically possible for index and db to be out of sync.
New behaviour: indexables are set to "indexOk=null", and _after_ the transaction on the db has gone through, the LuceneActor will find the indexable and update the index safely.