Fixed #55 - imageService croaking on not finding a file.
In the rare case that an image file is missing from the file system (which probably only happens when a bug prevents deletion of an OSD *after* the file it points to has been removed... which
would cause a database rollback, but cannot recover the deleted file), the imageService would try to read the non-existent file and then crash,
preventing the webclient from showing the folder which contains the problematic object.
The current code will create a 1x1 placeholder image so at least the folder can be visited again in the web client. Further fixing should report the broken object to the user.